Repos/Modeling/requirements.txt
numpy==1.16.6
pandas==0.24.2
sklearn==0.0
scikit-learn==0.20.4
catboost==0.23.2
imbalanced-learn==0.4.3
lightgbm==2.3.1
xgboost==0.80
PyHive==0.6.1
findspark==1.3.0
pyspark==2.4.4
scipy==1.2.3
PyYAML==5.3
seaborn==0.9.1
matplotlib==2.2.5
scikit-plot==0.3.7